+++ src/driver/sqlite-abstract/AbstractSqliteDriver.ts
@@ -12,7 +12,6 @@ import {TableColumn} from "../../schema-builder/table/TableColumn";
 import {BaseConnectionOptions} from "../../connection/BaseConnectionOptions";
 import {EntityMetadata} from "../../metadata/EntityMetadata";
 import {OrmUtils} from "../../util/OrmUtils";
-import {ArrayParameter} from "../../query-builder/ArrayParameter";
 
 /**
  * Organizes communication with sqlite DBMS.
@@ -275,10 +274,18 @@ export abstract class AbstractSqliteDriver implements Driver {
         if (!parameters || !Object.keys(parameters).length)
             return [sql, builtParameters];
 
-        const keys = Object.keys(parameters).map(parameter => "(:" + parameter + "\\b)").join("|");
+        const keys = Object.keys(parameters).map(parameter => "(:(\\.\\.\\.)?" + parameter + "\\b)").join("|");
         sql = sql.replace(new RegExp(keys, "g"), (key: string): string => {
-            let value = parameters[key.substr(1)];
-            if (value instanceof Array) {
+            let value: any;
+            let isArray = false;
+            if (key.substr(0, 4) === ":...") {
+                isArray = true;
+                value = parameters[key.substr(4)];
+            } else {
+                value = parameters[key.substr(1)];
+            }
+
+            if (isArray) {
                 return value.map((v: any) => {
                     builtParameters.push(v);
                     return "?";
@@ -289,7 +296,6 @@ export abstract class AbstractSqliteDriver implements Driver {
                 return value();
 
             } else {
-                if (value instanceof ArrayParameter) value = value.value;
                 builtParameters.push(value);
                 return "?";
                 // return "$" + builtParameters.length;